fix minor compiler warnings
git-svn-id: svn+ssh://svn.abisource.com/svnroot/link-grammar/trunk@24435 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ link-grammar/api-structures.h
@@ -42,10 +42,10 @@ struct Cost_Model_s
 typedef struct Resources_s * Resources;
 struct Resources_s
 {
-	int    max_parse_time;  /* was double before --DS */
-	int    max_memory;
+	int    max_parse_time;
+	size_t max_memory;
 	double time_when_parse_started;
-	int    space_when_parse_started;
+	size_t space_when_parse_started;
 	double when_created;
 	double when_last_called;
 	double cumulative_time;
